+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.2.4] - 2024-07-03
+
+### Bug Fixes
+- Wrong parenthesis on labels
+- Diagonal line on mixed-zoom plot
+
 ## [1.2.3] - 2024-07-01
 
 ### Changed